requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
discord.py==1.7.3
fastapi==0.73.0
frozenlist==1.3.0
h11==0.13.0
idna==3.3
multidict==6.0.2
pydantic==1.9.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.4
yarl==1.7.2